revert the tests to ensure backwards compatibility
